pyproject.toml
[project]
name = "anishathalye"
version = "0.1.0"
description = "GitHub README"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pygithub>=2.6.1,<3",
    "pyyaml>=6.0.2,<7",
]

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "ruff>=0.12.0",
    "types-pyyaml>=6.0.12.20250516",
]

[tool.mypy]
strict = true

[tool.ruff]
line-length = 120
lint.select = [
    # pycodestyle
    "E",
    "W",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # ruff-specific rules
    "RUF",
]
